Implement most command line options
Based on the homework3, this version implemented -d, -h, -i and -p
options. In addition, the server has already supported concurrent
access and can run as a daemon process.

Meanwhile, implement a String library which is in jstring.c. This
library provides a series of functions to manipulate strings like
Java. For more information, see jstring.c and jstring.h.