Implement logging function
Create a set_logging structure in http.h. This structure contains
information for logging. Implement logging function and clean_logging
function. Logging will take a opened file descriptor by O_APPEND. In
debug mode, fd will be STDOUT_FILENO. The function will check the
logging structure and return as an error if the values in structure
are illegal.

The function will return the number of bytes it writes to the logging
file. Return 0 if there is an error happend.