Change split string function and add new functions.
Add function check_version to test if the http_version is the same format
as 1*DIGIT"."1*DIGIT. And change the http_version in the http_request
structure to float.

Add a new split_str to get each line of the http requests. If the http
request is a bad request, split_str will return NULL. And it will
return "" if at the end of the request header field. This function is
designed as the same function as strtok_r. But strtok_r cannot split
by "\r\n". The new function is designed for this purpose.